refactor(fsdp): dispatch patches by Torch version
Keep the public policy module independent of private Torch internals and load the source-guarded 2.11 implementation only after matching the installed base version.

Authored with Claude.
