+++ tests/fast/backends/fsdp_utils/test_fsdp_param_dtype_patch.py
@@ -1,3 +1,5 @@
+import sys
+
 from tests.ci.ci_register import register_cpu_ci
 
 register_cpu_ci(est_time=5, suite="stage-a-cpu", labels=["fsdp"])
@@ -9,20 +11,22 @@ from miles.backends.fsdp_utils import fsdp_param_dtype_patch
 
 
 def test_patch_rejects_unpinned_torch(monkeypatch):
-    monkeypatch.delattr(
-        fsdp_param_dtype_patch._fsdp_collectives,
-        fsdp_param_dtype_patch._PATCH_SENTINEL,
-        raising=False,
+    implementation_module = (
+        "miles.backends.fsdp_utils._fsdp_param_dtype_patch_2_11"
     )
+    monkeypatch.delitem(sys.modules, implementation_module, raising=False)
     monkeypatch.setattr(torch, "__version__", "2.12.0+cu130")
 
-    with pytest.raises(RuntimeError, match="requires torch==2.11.0"):
+    with pytest.raises(RuntimeError, match="supports torch==2.11.0"):
         fsdp_param_dtype_patch.apply_param_dtype_map_patch()
+    assert implementation_module not in sys.modules
 
 
 def test_patch_rejects_source_drift(monkeypatch):
+    from miles.backends.fsdp_utils import _fsdp_param_dtype_patch_2_11
+
     monkeypatch.setitem(
-        fsdp_param_dtype_patch._SOURCE_HASHES,
+        _fsdp_param_dtype_patch_2_11._SOURCE_HASHES,
         "FSDPParamGroup.__init__",
         "0" * 64,
     )
@@ -31,9 +35,9 @@ def test_patch_rejects_source_drift(monkeypatch):
         RuntimeError,
         match="source hash for FSDPParamGroup.__init__ changed",
     ):
-        fsdp_param_dtype_patch._verify_source(
+        _fsdp_param_dtype_patch_2_11._verify_source(
             "FSDPParamGroup.__init__",
-            fsdp_param_dtype_patch._ORIGINAL_PARAM_GROUP_INIT,
+            _fsdp_param_dtype_patch_2_11._ORIGINAL_PARAM_GROUP_INIT,
         )
